Move configs to package instead of the same file and corrected timeout
